Remove duplicate error.
Calling req.abort() after a timeout causes the socket to be closed,
which emits an error on the request. Emitting another error causes the
callback to be invoked twice.

This has an especially bad interaction given the retry behavior - for
each failure this code would retry twice. Multiple failures would cause
retries to cascade rapidly.